rename entrypoint file
